fix(state): Verify the state inclusion proof without splitting keys (#1483)
## Overview

There has to be better way of fixing this, but I'm unsure of how we want
to handle it, so submitting this for now.

The bug is that, when verifying state inclusion proofs, tendermint will
perform [some
processing](https://github.com/celestiaorg/celestia-core/blob/a0ccbd068927b638705f63acdf6303d3091da985/crypto/merkle/proof_key_path.go#L85-L110)
on the keypath that you give it that the sdk simply doesn't have to do.
The iavl doesn't do this, as we just pass it the key it wants, not the
path. It just sticks the key that you give in the [request right
back](https://github.com/cosmos/cosmos-sdk/blob/b46d83850270b82496e4cd61421d9ebc8bac87f2/store/iavl/store.go#L324-L326),
and because the processing in tendermint sometimes results in a
different key, the proof verification
[fails](https://github.com/celestiaorg/celestia-core/blob/a0ccbd068927b638705f63acdf6303d3091da985/crypto/merkle/proof_op.go#L52).

Specifically, the string encoding of the bytes we pass it sometimes have
the string `/` and `%` in them from the address, the processing
tendermint does thinks its supposed to do something when it actually
isn't, and then the correct key returned by the sdk no longer matches.
This is why we're only seeing some addresses break!

[bad processing point
1](https://github.com/celestiaorg/celestia-core/blob/a0ccbd068927b638705f63acdf6303d3091da985/crypto/merkle/proof_key_path.go#L91)
[bad processing point
2](https://github.com/celestiaorg/celestia-core/blob/a0ccbd068927b638705f63acdf6303d3091da985/crypto/merkle/proof_key_path.go#L102)
(idek why this is there)

My quick and dirty solution was to copy paste fork ~~the file from
tendermint~~ a few methods that verify proofs, then change the api so we
don't have to split the string :laughing: . We should fix it properly
soon. This might be a bug with tendermint, but I was admittedly too
tired to look into it or write up an issue. We might just need to pass a
different formatted string(?)

as a bonus, I added the foundation for an integration test for the core
accessor, that should make it waaaaaaaayyyy easier to round trip test
things! :slightly_smiling_face: :christmas_tree:

pls feel to edit this PR and add whatever else is needed to get this
across the finish line

closes #1461 

## Checklist

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords

Co-authored-by: rene <41963722+renaynay@users.noreply.github.com>