fix(header/p2p): preconnect trusted peers on header-ex start (#1552)
We need to connect to trusted peers and kick off various protocols to start talking to them. In the past, a node was always connected to trusted peers to get headers from them. As from recent improvements to header-ex, we now only rely on trusted peers for initialization. Hence, if a node is already initialized, it won't contact go to connect to trusted peers, and this commit fixes this.

This was found during fixing no trusted peers issue in swamp tests in #1556