feat!: add versioning for pubsub topics in fraud, header packages (#1620)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview
Resolves https://github.com/celestiaorg/celestia-node/issues/1455,
https://github.com/celestiaorg/celestia-node/issues/1454,
https://github.com/celestiaorg/celestia-node/issues/1453

PLEASE NOTE: this PR is a protocol breaking since it introduces changes
in protocol namings.
<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords