refactor(das): Don't increment totalSampled if header `isRecent` (#1736)
This (not very elegantly) fixes an issue with tracking `totalSampled`
inside the DASer which results in 2x the amount of sampled headers as
were actually sampled.

From @walldiss: 
```
It happens because DASer does not call Sampling directly, instead it calls SharesAvailable that ensures that block is sampled. Under the hood, SharesAvailable will check the internal cache and if it hits, it will return immediately. This logic was utilised in DASer design, to simplify independent tracking of recent headers sampling. So sampling recent headers could call SharesAvailable twice: first by recent sampling job and by catchup job later. It was done to simplify already complex DASer logic.
```

Eventually, `totalSampled` should probably be tracked inside
`CacheAvailability` as that's the only place that will add an entry for
a height one time only.