das/share: Change log from DAH.Hash() to DAH.String() (#1971)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

Changes log input to use the string method for both the
DataAvailabilityHeader and Root.

Fixes #1898

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->


- [x] Required CI checks are passing
- [x] Linked issues closed with keywords