feat(daser): introduce retry logic for daser (#1860)
## Overview
DASer jobs now have `job_type` field that identifies purpose of the job.
There are 3 possible types of jobs:

 - catchup
 - recent
 - retry

Introduce new type of jobs for DASer - `retry`. `retry` jobs will be
launched only if there is enough concurrency slots after running
`recent` jobs and `catchup` jobs. Meaning DASer will start retry with
lowest priority, only when catchup is done and recent header are synced
without falling behind.

Also includes minor simplifications for affected code.

Resolves https://github.com/celestiaorg/celestia-node/issues/1782,
https://github.com/celestiaorg/celestia-node/issues/2027
<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [ ] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [x] Linked issues closed with keywords

---------

Co-authored-by: rene <41963722+renaynay@users.noreply.github.com>
Co-authored-by: Ryan <ryan@celestia.org>
Co-authored-by: Hlib Kanunnikov <hlibwondertan@gmail.com>