refactor(modheader): use WithParams options (#2102)
Just removes the need to pass every particular option and pass the whole config instead.

The main motivation for this PR is to minimize cases where we add a new config field but forget to add and set the option for it. This has happened with TrustedPeersRequestTimeout, and made a new waste a ton of time debugging in the wrong direction as I assumed the configuration value was set correctly.