fix(state): JSON marshaling for sdk Address wrapper (#2348)
While debugging #2322 , I noticed that it is not only an RPC issue, but
also a broken API one. All RPC methods that used the type
`state.Address` were unusable, both via the client and via raw JSON
calls. This is because the server was unable to marshal the address
string value back into the interface type.

To circumvent this, I have embedded the sdk.Address type in the same way
that we embed the fraud proof type, to allow us to unmarshal it back
into a concrete type. I have also added unit tests to fix this.

In addition, it fixes the RPC parsing - so it closes #2322 .