feat(header): Check app version in eh validation and return err on mismatch (#2138)
As recommended by @cmwaters, we should ensure (as a rudimentary check)
that the headers being passed to us from core match in app versions as
upgraded app versions can indicate breaking changes.

We should error out if the app version does not match.

This PR is still draft until celestia-app has a version out where they
export the AppVersion as a const that we can depend on.

Eventually we should support several versions via #2137 but for now,
this is a good enough check.

TODO: 
- [x] depend on const from app
- [x] test

---------

Co-authored-by: Hlib Kanunnikov <hlibwondertan@gmail.com>