fix(blob): typo, sumitter -> submitter (#2648)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.
-->

## Overview

- simple change in `blob/service.go` where a typo of sumitter existed in
the `Service` struct, renamed to submitter

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 
-->

## Checklist

<!-- 
Please complete the checklist to ensure that the PR is ready to be
reviewed.

IMPORTANT:
PRs should be left in Draft until the below checklist is completed.
-->

- [√] New and updated code has appropriate documentation
- [√] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [n/a] Visual proof for any user facing features like CLI or
documentation updates
- [n/a] Linked issues closed with keywords

---------

Co-authored-by: Ryan <ryan@celestia.org>