feat(share/eds): Rework accessor cache (#2612)
**Motivation** 
Previously, all accessors went into cache. They were closed only when
they were evicted from it and there was no way to manually close one.
Now, when we need to cache accessors only in some cases and in some not
to, this approach will not work. Enabling control of accessor close
required major cache rework.

New cache features:
- Old cache didn't allow to close and release of an accessor. Now,
Instead of returning just a reader/blockstore, cache returns object that
composed of builder and `io.Closer` (Accessor interface). This allows
caller to have an ability to close and release accessor once it is done
with reading from it.
- Items stored in cache also provide io.Closer`, but it is noop.
Lifecycle of cached items is controlled by cache itself. For now objects
are released right after they are evicted from cache, but it could be
changed to smarter release based on current readers count. Will be
implemented as separate issue.
- Lazy blockstore builder. Old cache created blockstore every time
accessor was acquired, even when only reader was used and blockstore was
not needed. Creating blockstore is costly operation that involves
building index, so this was reworked to create blockstore, only when it
is requested and store it for subsequent requests.
- adds cache unit tests.
- adds metrics for cache methods to track cache hits/miss.
- general cleanup of unused methods/logic.

It will be easier for reviewer to start from `store.go` and
`blockstore.go` and then go into new `cache` pkg