chore: Update docker-build-publish.yml (#2838)
This fixes the bug that rollkit local-celestia-devent was experiencing with the docker images being overwritten.

The changes to the reuseable docker action simplify the logic for easier maintenance.

The changes to the node CI file are to only run the docker action on a PR or pushes to main. When the workflow was triggering on all push events, the action was running twice on PRs which was contributing to the confusion and causing images to be overwritten.

This means if the team wants a docker image built for testing, they must open at least a draft PR. The PR doesn't need to be merged, but the PR trigger is what will built the image.