test(gateway): slight refactor of some presentation and add some endpoint registration tests (#2939)
goals: readability, test coverage

Appreciate that the gateway is v unimportant but is exists. I was
starting at the top just looking for places to give a clean up and add
some unit test coverage and found this.

Mostly trying to make it cleaner to read, then added some tests, also
wanted to add a `/status/health` endpoint just for ease, i would want
this as a node operator if i was running with `--gateway` for whatever
reason

Also took liberty to take some error handling code out of `writeError`
that would never execute. In trying to write a test to trigger the
json.Marshal error case, i realized that as err.Error() always returns a
string it could never trigger json.Unmarshal to fail without panic
before even getting there, so removed that code path.

i might revisit this and give it another refactor at some point soon,
unless we're planning to remove this completely?

on to next area